moved SxAnimalsSubparser to __init__ file
