+++ src/sxapi/cli/parser/subparser/__init__.py
@@ -1,2 +1,2 @@
-from sxapi.cli.parser.subparser.animals.animals import SxApiAnimalsSubparser
+from sxapi.cli.parser.subparser.animals import SxApiAnimalsSubparser
 from sxapi.cli.parser.subparser.token import SxApiTokenSubparser
